+++ lib/redix.ex
@@ -446,6 +446,19 @@ defmodule Redix do
   end
 
   @doc """
+  Subscribes `recipient` to the given channel or list of channels, raising if
+  there's an error.
+
+  Works exactly like `subscribe/4` but raises if there's an error.
+  """
+  @spec subscribe!(GenServer.server, String.t | [String.t], pubsub_recipient, Keyword.t) :: :ok
+  def subscribe!(conn, channels, recipient, opts \\ []) do
+    conn
+    |> subscribe(channels, recipient, opts)
+    |> raise_for_pubsub_errors
+  end
+
+  @doc """
   Subscribes `recipient` to the given pattern or list of patterns.
 
   Works like `subscribe/3` but subscribing `recipient` to a pattern (or list of
@@ -465,6 +478,19 @@ defmodule Redix do
   def psubscribe(conn, patterns, recipient, opts \\ []) do
     timeout = opts[:timeout] || @default_timeout
     Connection.call(conn, {:psubscribe, List.wrap(patterns), recipient}, timeout)
+  end
+
+  @doc """
+  Subscribes `recipient` to the given pattern or list of patterns, raising if
+  there's an error.
+
+  Works exactly like `psubscribe/4` but raises if there's an error.
+  """
+  @spec psubscribe!(GenServer.server, String.t | [String.t], pubsub_recipient, Keyword.t) :: :ok
+  def psubscribe!(conn, patterns, recipient, opts \\ []) do
+    conn
+    |> psubscribe(patterns, recipient, opts)
+    |> raise_for_pubsub_errors
   end
 
   @doc """
@@ -490,6 +516,19 @@ defmodule Redix do
   end
 
   @doc """
+  Subscribes `recipient` to the given channel or list of channels, raising if
+  there's an error.
+
+  Works exactly like `unsubscribe/4` but raises if there's an error.
+  """
+  @spec unsubscribe!(GenServer.server, String.t | [String.t], pubsub_recipient, Keyword.t) :: :ok
+  def unsubscribe!(conn, channels, recipient, opts \\ []) do
+    conn
+    |> unsubscribe(channels, recipient, opts)
+    |> raise_for_pubsub_errors
+  end
+
+  @doc """
   Unsubscribes `recipient` from the given pattern or list of patterns.
 
   This function basically "undoes" what `psubscribe/3` does: it unsubscribes
@@ -509,6 +548,19 @@ defmodule Redix do
   def punsubscribe(conn, patterns, recipient, opts \\ []) do
     timeout = opts[:timeout] || @default_timeout
     Connection.call(conn, {:punsubscribe, List.wrap(patterns), recipient}, timeout)
+  end
+
+  @doc """
+  Subscribes `recipient` to the given pattern or list of patterns, raising if
+  there's an error.
+
+  Works exactly like `punsubscribe/4` but raises if there's an error.
+  """
+  @spec punsubscribe!(GenServer.server, String.t | [String.t], pubsub_recipient, Keyword.t) :: :ok
+  def punsubscribe!(conn, patterns, recipient, opts \\ []) do
+    conn
+    |> punsubscribe(patterns, recipient, opts)
+    |> raise_for_pubsub_errors
   end
 
   @doc """
@@ -540,4 +592,11 @@ defmodule Redix do
         val
     end
   end
+
+  defp raise_for_pubsub_errors(:ok),
+    do: :ok
+  defp raise_for_pubsub_errors({:error, %Redix.Error{} = err}),
+    do: raise(err)
+  defp raise_for_pubsub_errors({:error, err}),
+    do: raise(Redix.ConnectionError, err)
 end