+++ CHANGELOG.md
@@ -6,9 +6,13 @@
   and more flexible).
 * Add support for PubSub. The following functions have been added to the `Redix` module:
   * `Redix.subscribe/4`
+  * `Redix.subscribe!/4`
   * `Redix.psubscribe/4`
+  * `Redix.psubscribe!/4`
   * `Redix.unsubscribe/4`
+  * `Redix.unsubscribe!/4`
   * `Redix.punsubscribe/4`
+  * `Redix.punsubscribe!/4`
   * `Redix.pubsub?/2`
 
 ## v0.1.0